replacement: Converted ReplacementProvider to a Service
