+++ src/main/java/com/ystrazhko/git/services/ServiceProvider.java
@@ -30,5 +30,6 @@ public class ServiceProvider {
                 new GroupsUserServiceImpl(new RESTConnectorFactory().getRESTConnector()));
         _services.put(ProjectService.class.getName(),
                 new ProjectServiceImpl(new RESTConnectorFactory().getRESTConnector()));
+        _services.put(ReplacementService.class.getName(), new ReplacementServiceImpl());
     }
 }
